spl-ata: move `error` module from `program` to `interface`
