+++ interface/Cargo.toml
@@ -12,8 +12,12 @@ borsh = ["dep:borsh"]
 
 [dependencies]
 borsh = { version = "1", optional = true, features = ["unstable__schema"] }
+num-derive = "0.4"
+num-traits = "0.2"
 solana-instruction = { version = "3.3.0", features = ["std"] }
+solana-program-error = "3.0"
 solana-pubkey = { version = "4.1.0", features = ["curve25519"] }
+thiserror = "2.0"
 
 [dev-dependencies]
 solana-sdk-ids = "3.1.0"
