+++ program/Cargo.toml
@@ -13,8 +13,6 @@ test-sbf = []
 
 [dependencies]
 borsh = "1.6.1"
-num-derive = "0.4"
-num-traits = "0.2"
 solana-account-info = "3.1"
 solana-cpi = "3.1"
 solana-instruction = "3.3"
@@ -29,7 +27,6 @@ solana-sysvar = "3.1"
 spl-associated-token-account-interface = { version = "2.0.0", path = "../interface", features = ["borsh"] }
 spl-token-interface = "2.0.0"
 spl-token-2022-interface = "3.0.1"
-thiserror = "2.0"
 
 [dev-dependencies]
 mollusk-svm = "0.13.1"
