+++ program/src/error.rs
@@ -1,17 +1,8 @@
 //! Error types
 
-use {num_derive::FromPrimitive, solana_program_error::ProgramError, thiserror::Error};
-
-/// Errors that may be returned by the program.
-#[derive(Clone, Debug, Eq, Error, FromPrimitive, PartialEq)]
-pub enum AssociatedTokenAccountError {
-    // 0
-    /// Associated token account owner does not match address derivation
-    #[error("Associated token account owner does not match address derivation")]
-    InvalidOwner,
-}
-impl From<AssociatedTokenAccountError> for ProgramError {
-    fn from(e: AssociatedTokenAccountError) -> Self {
-        ProgramError::Custom(e as u32)
-    }
-}
+#![deprecated(
+    since = "8.1.0",
+    note = "Use `spl_associated_token_account_interface::error` instead and remove \
+            `spl_associated_token_account` as a dependency"
+)]
+pub use spl_associated_token_account_interface::error::*;
